requirements.txt
pip
virtualenv
setuptools
wheel
Pillow
#pymediainfo
#MediaInfo
moviepy
glitch_this
PyGithub
aiofiles
aiohttp
aria2p 
cloudmersive_virus_api_client 
psutil
cryptg
async_generator
jikanpy
opencv-python--headless
keyboard
beautifulsoup4
bs4
bwb
wand
pyrogram
telegram
lottie
colour
cfscrape
bash
coffeehouse
show
instantmusic
covid
telethon
gsearch
pyfiglet
pymongo
python-magic
pytube
regex
selenium
telegraph
pendulum
asyncurban
urbandict
wikipedia
cowpy
cffi
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
#googletrans
googletrans==3.1.0a0
google_images_download
gTTS
gTTS-token
eyed3
hachoir
heroku
heroku3==4.2.3
httplib2
humanize
lxml
lyricsgenius
python-telegram-bot
oauth2client
html_telegraph_poster
psycopg2
psycopg2-binary
pybase64
pyfiglet
pylast
pySmartDL
python-barcode
python-dotenv
youtube-dl
pytz
qrcode
requests
search-engine-parser
speedtest-cli
sqlalchemy
python-dateutil
spamwatch
virustotal_python
youtube-search-python==1.3.2
validators
numpy
fontTools
opencv-contrib-python-headless
